Refctor: rename data folder from dataloading to just data
As said. I think the folder name is too long and not necessary. I think /src/data is pretty clear.
