+++ src/App.tsx
@@ -1,6 +1,6 @@
 import PageNavBar from './controls/PageNavBar';
 import { PageParamsProvider } from './controls/PageParamsContext';
-import { DataProvider } from './dataloading/DataContext';
+import { DataProvider } from './data/DataContext';
 import Footer from './Footer';
 import { HoverCardProvider } from './generic/HoverCardContext';
 import MainViews from './views/MainViews';
@@ -16,9 +16,6 @@ function App() {
             <MainViews />
           </div>
           <ViewModal />
-          {/* <footer>
-            &copy; 2025 <a href="https://translationcommons.org/">Translation Commons</a>
-          </footer> */}
           <Footer />
         </HoverCardProvider>
       </DataProvider>
