+++ src/data/GlottologData.tsx
