+++ src/views/census/CensusHierarchy.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import { getScopeFilter } from '../../controls/filter';
 import { getSortFunction } from '../../controls/sort';
-import { useDataContext } from '../../dataloading/DataContext';
+import { useDataContext } from '../../data/DataContext';
 import { CensusData } from '../../types/CensusTypes';
 import { ObjectData, TerritoryData } from '../../types/DataTypes';
 import { ObjectType } from '../../types/PageParamTypes';
