+++ src/views/language/LanguageHierarchy.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import { getScopeFilter } from '../../controls/filter';
 import { usePageParams } from '../../controls/PageParamsContext';
 import { getSortFunction } from '../../controls/sort';
-import { useDataContext } from '../../dataloading/DataContext';
+import { useDataContext } from '../../data/DataContext';
 import { ObjectData } from '../../types/DataTypes';
 import { LanguageData, LanguageSchema, LanguageScope } from '../../types/LanguageTypes';
 import { ObjectType } from '../../types/PageParamTypes';
